+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2026-02-16
+
+### Added
+- A way to get current active elements
+- Dependency for uGUI 2.0.0
+
+### Changed
+- Made the UIManagerBase.OnDestroy method virtual
+
 ## [1.2.5] - 2025-01-10
 
 ### Added